Updated effect of luck on map generation, updated plots.
This makes high roll counts EXTREMELY time consuming,
but it fixes bugs in the accept/reject method. No luck
now has a map mean in the proper range, and a 'more linear'
effect of luck on the distribution.

Also added an updated plotting method. Decided to stay
with Gadfly, as Plots wasn't quite working for the large
(not that large?) datasets that I was generating.

Instead, the plotting method was changed from Geom.histogram
to Geom.histogram2d. This allows comparing the root and zoom
on the same plot. It clearly shows a comparison of the means.

Also fixed plot scaling, for easier comparison of distributions.