Added configuration options from file
Not completely happy with this implementation. The JSON file didn't
accept any newlines, which really hurts readability. It also doesn't
preserve typing on the arrays used in MAP_SIZE and IMAGE_SIZE, and won't
preserve typing on the LUCK parameter (requires conversion to UInt8
after JSON is loaded at runtime).

Might want to try to use some of Julia's code generation functionality
for configuration. Will require some thought and planning, but it should
lead to a pretty flexible, dynamically generated testing program. Might
not be the best solution, but it could be a good spot to test out some
Julia metaprogramming.