UNTESTED Implementation of graze function and helpers.
Goal of this function is to have a 'creature' that moves across the map,
consuming resources (currently moisture). After consuming resource on
the current tile, it should move to the highest valued tile in its
vicinity, and continue as such.

Current bugs include some out of bounds exceptions, and jumping outside
of the neighborhood. I think these are related.

The basic functionality is here, and it writes to a map image file for a
quantitative analysis of behavior, but rigorious tests are required now
for hunting down the bugs. This commit includes 4 new functions (2 are
pretty much the same), so my next move should be to write unit tests for
these (other unit tests can come after, I feel like the others will be
more difficult because of the randomness...)